requirements.txt
youtube_transcript_api
google-genai